remove unused code, support retina
