+++ iina/MainWindowController.swift
@@ -677,8 +677,6 @@ class MainWindowController: NSWindowController, NSWindowDelegate {
   // resize framebuffer in videoView after resizing.
   func windowDidEndLiveResize(_ notification: Notification) {
     videoView.videoSize = window!.convertToBacking(videoView.bounds).size
-    // new (empty) frame buffer is created, so draw a frame manually
-    // videoView.drawFrame()
   }
 
   // MARK: - Control UI
